+++ tools/importers/darknet/darknet_import.py
@@ -58,6 +58,7 @@ class DarknetImporter:
 if __name__ == "__main__":
     
     logging.basicConfig(level=logging.INFO, format="%(message)s")
+    logger.setup()
     
     parser = argparse.ArgumentParser(
         formatter_class=argparse.RawDescriptionHelpFormatter,