+++ tools/utilities/pythonlibs/audio/training/make_training_list.py
@@ -24,17 +24,49 @@ def is_background_noise(filename):
     return "_background_noise_" in filename
 
 
+def shuffle(filename, output):
+
+    if not os.path.isfile(filename):
+        print("Cannot find file: {}".format(filename))
+        return
+
+    with open(filename, "r") as f:
+        testlist = f.readlines()
+
+    testlist = np.random.choice(testlist, len(testlist), replace=False)
+
+    with open(output, "w") as f:
+        f.writelines(testlist)
+
+
 def make_training_list(wav_files, max_files_per_directory):
     """
     Create a training list file given the directory where the wav files are organized into subdirectories,
     with one subdirectory per keyword to be recognized.  This training list will exclude any files
-    already referenced by the 'testing_list.txt' or 'validation_list.txt'
+    already referenced by the 'testing_list.txt' or 'validation_list.txt'. It will also shuffle the
+    training list, and the testing_list and validation_list to ensure training and testing is fair.
     """
     if not os.path.isdir(wav_files):
         print("wav_file directory not found")
         return
 
-    ignore_list = load_list_file(os.path.join(wav_files, "testing_list.txt"))
+    testing_list = os.path.join(wav_files, "testing_list.txt")
+    if not os.path.isfile(testing_list):
+        print("### error: testing list '{}' file not found".format(testing_list))
+        return
+
+    validation_list = os.path.join(wav_files, "validation_list.txt")
+    if not os.path.isfile(validation_list):
+        print("### error: validation list '{}' file not found".format(validation_list))
+        return
+
+    print("Shuffling: {}".format(testing_list))
+    shuffle(testing_list, testing_list)
+
+    print("Shuffling: {}".format(validation_list))
+    shuffle(validation_list, validation_list)
+
+    ignore_list = load_list_file(testing_list)
     ignore_list += load_list_file(os.path.join(wav_files, "validation_list.txt"))
 
     list_file_name = os.path.join(wav_files, "training_list.txt")
@@ -61,11 +93,14 @@ def make_training_list(wav_files, max_files_per_directory):
                         file_list += [entry]
                         count += 1
 
-            if len(file_list) > max_files_per_directory:
-                file_list = np.random.choice(np.array(file_list), max_files_per_directory)
-                print("choosing {} random files from {}".format(len(file_list), dir_name))
-            else:
-                print("found {} files in {}".format(len(file_list), dir_name))
+            max = max_files_per_directory
+            if len(file_list) < max_files_per_directory:
+                max = len(file_list)
+
+            # shuffle the training list also
+            file_list = np.random.choice(np.array(file_list), max)
+            print("choosing {} random files from {}".format(len(file_list), dir_name))
+
             for e in file_list:
                 f.write(e + "\n")